requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
bcrypt==4.1.2
python-jose[cryptography]
python-multipart==0.0.6
email-validator==2.1.0

